Refactor App logic into reusable hooks
