+++ README.md
@@ -106,5 +106,6 @@ A lo largo del proyecto se han añadido numerosas mejoras, entre ellas:
 - Buscador integrado para localizar "Espada" o "Daga" y agregarlas a la biblioteca.
 - Corregido un error de animación al aumentar la cantidad de un objeto en el inventario.
 - Si aparece un error de ESLint sobre `prettier`, ejecuta `npm install` para instalar todas las dependencias.
+- Se introdujeron hooks personalizados (`useLogin`, `useGlossary`, `useResources`) para dividir la lógica de `App.js` en módulos reutilizables.
